backend/requirements.txt
# REQUIREMENTS

flask
flask-cors
openai
python-dotenv
requests
cohere
